requirements.txt
openai>=1.0.0
rich>=10.0.0
pyperclip>=1.8.0
pyyaml>=6.0.0
aiohttp>=3.8.0
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
coverage>=7.0.0
mypy>=1.0.0
mcp>=0.1.0
anyio>=4.0.0
click>=8.0.0
starlette>=0.27.0
uvicorn>=0.24.0
httpx>=0.24.0
